Implemented functionality to produce a string from a file of one line.

+++ file_reader.h
@@ -5,7 +5,7 @@ long file_length(const char *filename);
 
 void file_to_string(
 	const char *filename,
-	const int file_length,
+	const long file_length,
 	char *string);
 
 #endif
